(fix) #46  Make tests pass again (by moving test contracts into the test fixture) (feat) #46 Refactor proxy injection out of the deployer
