+++ test/fixture-projects/hardhat-project/contracts/scilla/ADTTest.scilla
