+++ test/fixture-projects/hardhat-project/contracts/scilla/AdditionLib.scillib
