+++ test/fixture-projects/hardhat-project/contracts/scilla/Codehash.scilla
