+++ test/fixture-projects/hardhat-project/contracts/scilla/GenerateAdtType.scilla
