+++ test/fixture-projects/hardhat-project/contracts/scilla/HelloWorld.scilla
@@ -30,6 +30,3 @@ transition getHello ()
     e = {_eventname: "getHello()"; msg: r};
     event e
 end
-transition throwError ()
-    throw
-end
