+++ test/zilliqa.live.test.ts
@@ -38,7 +38,6 @@ describe("", function () {
       const privKey = this.zobj.createPrivateKey();
       const [acc,idx] = this.zobj.pushPrivateKey(privKey);
       const VAL = new BN("100000000000000001000");
-      console.log(`Transferring ${VAL} to ${acc}`);
       const txn = await this.zobj.transferTo(acc, new BN(VAL));
       const [bal,nonce] = await this.zobj.getBalance(acc);
       const transferredBalance = bal;
@@ -48,7 +47,6 @@ describe("", function () {
       // Now transfer it back.
       this.hre.setActiveAccount(idx);
       // Lose 10 zil here for gas.
-      console.log(`Transferredbalance ${transferredBalance}`);
       const txn2 = await this.zobj.transferToAddress(this.zobj.getAccounts()[0].address, new BN(1_000));
       expect(txn2['receipt']['success']).to.be.true;
       this.hre.setActiveAccount(0);
